Draft: Add optional `no_std` support
This is a followup on a related commit to the `special` crate at
<https://github.com/robamler/special/commit/adf6a15> that propagates
functionality up the dependency hierarchy.

This commit should make it possible to compile `probability` for targets
that don't have a system `libm` C library, such as microcontrollers and
WebAssembly. This is achieved in a backward compatible way by adding a
cargo feature "std" that is activated by default. As long as a crate
that depends on `probability` doesn't explicitly deactivate the "std"
feature, no behavior changes. But if a crate that uses `probability`
deactivates the "std" feature (see blow), then `probability` now
forwards this to its dependency on the `special` crate, which will cause
`special` to replace its use of the system `libm` C library with a
reimplementation in pure rust.

In addition, this commit adds the line
`#![cfg_attr(not(feature = "std"), no_std)]`
to `lib.rs`, which makes it possible to use `probability` in `no_std`
environments (e.g., microcontrollers) if the "std" feature is
deactivated. Again, nothing changes as long as the "std" feature is
active, which is the default.

An additional benefit of (optionally) removing an FFI dependency is that
one can now use miri to test for undefined behavior in `probability` and
in crates that depend on `probability`. To do so, run:
`cargo +nightly miri test --no-default-features`

Unfortunately, benchmarks indicate that the `libm` rust crate is
slightly slower than the `libm` C library, which is why this commit
still uses the system `libm` C library as long as the "std" feature is
activated (which is the default).